Split Ultralight from North and South bound
